Renombramiento y movilizacion de archivos. Se integra helper de entidades, clase padre de characters, animaciones y ataque en la seccion de ENTITY. Asi mismo, para la seccion de map se integra la clase Paths unificada en una sola clase y documentada. Por ende, se actualizan imports en todos los archivos que sufren efectos de los cambios realizados.
