+++ src/entity/kinematic.py
